Rearrange the Upload CTA stuff
