+++ apps/Test App/Test App/ContentView.swift
@@ -13,7 +13,7 @@ struct ContentView: View {
     
     var body: some View {
         NavigationView {
-            //CreateUploadScreen()
+            CreateUploadScreen()
         }.preferredColorScheme(.dark)
     }
 }
