+++ apps/Test App/Test App/Screens/UploadListScreen.swift
@@ -1,0 +1,20 @@
+//
+//  UploadListScreen.swift
+//  Test App
+//
+//  Created by Emily Dixon on 5/15/23.
+//
+
+import SwiftUI
+
+struct UploadListScreen: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct UploadListScreen_Previews: PreviewProvider {
+    static var previews: some View {
+        UploadListScreen()
+    }
+}
